- Eliminate warnings in debug print macros by explicitly converting all   field to unsigned long.
Former-commit-id: a378a96e1f288035e57e7e1dc4c0152d3324029e

+++ sys/gnu/fs/ext2fs/ext2_vfsops.c
@@ -1798,7 +1798,7 @@ name|V
 parameter_list|(
 name|v
 parameter_list|)
-value|printf(#v"= %d\n", fs->v);
+value|printf(#v"= %lu\n", (unsigned long)fs->v);
 endif|#
 directive|endif
 name|fs
